% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{choose.cross.validate.low.rank.plus.sparse}
\alias{choose.cross.validate.low.rank.plus.sparse}
\title{Select a model based on the outcome of a K-fold cross-validation}
\usage{
choose.cross.validate.low.rank.plus.sparse(xval.path, method = "min")
}
\description{
Selects the best model according to cross-validation. If the method is "min"
return the fit with the smallest average cross-validated log-likelihood.
If the method is "hastie", return the most parsimonious fit whose cross-validated
log-likelihood is within one standard deviation of the minimum.
}
