% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lscggm.R
\name{.compute_proximal_operator}
\alias{.compute_proximal_operator}
\title{Compute proximal operator}
\usage{
.compute_proximal_operator(X, p, q, rho, tol, maxiter)
}
\description{
Compute the proximal operator for one step of the LSCGGM algorithm
}
