% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_path.R
\name{fit.low.rank.plus.sparse.path}
\alias{fit.low.rank.plus.sparse.path}
\title{Fit a path of the LRpS}
\usage{
fit.low.rank.plus.sparse.path(Sigma, gamma, n, lambdas = NULL,
  lambda.max = NULL, lambda.ratio = 1e-04, n.lambdas = 20,
  max.sparsity = 0.5, max.rank = NA, tol = 1e-05, max.iter = 2000,
  mu = 0.1, verbose = FALSE)
}
\description{
Fit a path of the LRpS
}
