% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrps.R
\name{fit.low.rank.plus.sparse}
\alias{fit.low.rank.plus.sparse}
\title{Fit the the low-rank plus sparse estimator}
\usage{
fit.low.rank.plus.sparse(Sigma, Lambda1, Lambda2, n, init = NULL,
  maxiter = 2000, mu = 0.1, tol = 1e-05, eta = 0.999,
  print_progress = T, print_every = 20, zeros = NULL,
  max.rank = NA)
}
\arguments{
\item{Sigma}{An estimate of the correlation matrix.}

\item{Lambda1}{Penalty on the l1 norm of S}

\item{Lambda2}{Penalty on the sum of the eigenvalues of L}

\item{n}{Number of samples}

\item{init}{The output of a previous run of the algorithm. For warm starts.}

\item{maxiter}{Maximal number of iterations}

\item{mu}{Stepsize of the ADMM algorithm}

\item{tol}{Relative tolerance required to stop the algorithm. Algorithm is stopped
when either the relative change in log-likelihood is below this threshold. or the
relative change in the Frobenius norm of the parameters is below this threshold.}

\item{eta}{Restart parameter in the accelerated ADMM}

\item{print_progress}{Whether the algorithm should report on its progress}

\item{print_every}{How often should the algorithm report on its progress (in terms
of #iterations).}

\item{zeros}{A p x p matrix with entries set to 0 or 1. Whereever its entries are
0, the entries of the estimated S will be set to 0.}

\item{max.rank}{The maximum rank of the estimated L. Especially useful if n > p.
By default it is min(p-1, n-1).}
}
\description{
Fit the the low-rank plus sparse estimator
}
\details{
Given a n x p data matrix X and its empirical correlation matrix
\eqn{\Sigma}, an alternative direction method of multipliers (ADMM) algorithm
is used to obtain the solutions \eqn{S, L} to
\deqn{(S, L) = argmin_{A, B}  -logdet(A - B) + Tr((A-B)\Sigma) + \lambda_1 ||A||_1 + \lambda_2Tr(B),}
subject to \eqn{A - B > 0} and \eqn{B \ge 0}.
}
