% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{generate.data}
\alias{generate.data}
\title{Generate synthetic data}
\usage{
generate.data(n, p, h, sparsity = 0.02, sparsity.latent = 0.7,
  outlier.fraction = 0)
}
\arguments{
\item{n}{Number of samples.}

\item{p}{Number of observed variables.}

\item{h}{Number of hidden variables.}

\item{sparsity}{Real between 0 and 1. The density of the sparse graph.}

\item{sparsity.latent}{Real between 0 and 1. Probability of connection between any
observed variable and any observed hidden variable.}

\item{outlier.fraction}{Fraction of the samples that should be drawn from a Cauchy
distribution. The remaining ones are drawn from a multivariate normal with the same
scale matrix.}
}
\value{
A list with keys:
 - obs.data: n x p data matrix of observed variables
 - full.data: (n+h) x p data matrix which also contains the hidden variables
 - true.precision.matrix: (n+h)x(n+h) matrix from which the data was sampled.
}
\description{
Generate data from Gaussian graphical model with hidden variables.
}
