% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lscggm_path.R
\name{lscggmadmm.path}
\alias{lscggmadmm.path}
\title{Compute the conditional low-rank plus sparse estimator (LSCGGM) along a path (for a fixed value of \eqn{gamma})}
\usage{
lscggmadmm.path(
  SigmaZ,
  SigmaZX,
  SigmaX,
  gamma,
  lambdas = NULL,
  lambda.max = NULL,
  lambda.ratio = 1e-04,
  n.lambdas = 20,
  max.sparsity = 0.5,
  max.rank = NA,
  rel_tol = 0.01,
  abs_tol = 1e-04,
  max.iter = 2000,
  mu = 1,
  verbose = FALSE
)
}
\arguments{
\item{SigmaZ}{An estimate of the correlation matrix of Z. m x m matrix.}

\item{SigmaZX}{An estimate of the cross-correlation matrix between Z and X. p x m matrix.}

\item{SigmaX}{An estimate of the correlation matrix of X. p x p matrix.}

\item{gamma}{A real between 0 and 1. The value of the tuning parameter gamma in the
parametrisation of the penalty described above. This is the trade-off between the sparse and trace penalties.}

\item{lambdas}{A decreasing sequence of values of lambda. See Details for the default value.}

\item{lambda.max}{A positive real. Maximum value of lambda. See Details.}

\item{lambda.ratio}{A real between 0 and 1. The smallest value of lambda is given by lambda.max * lambda.ratio. See Details.}

\item{n.lambdas}{A positive integer. The number of values of lambda to
generate according a geometric sequence between lambda.max and lambda.max * lambda.ratio. See Details.}

\item{max.sparsity}{A real between 0 and 1. Abort the computation of the path if S becomes denser than this value.}

\item{max.rank}{A real between 0 and 1. Abort the computuation of the path if the rank of L becomes higher than this value.}

\item{rel_tol}{\code{rel_tol} parameter of the \code{lrpsadmm} function.}

\item{abs_tol}{\code{rel_tol} parameter of the \code{lrpsadmm} function.}

\item{max.iter}{\code{max.iter} parameter of the \code{lrpsadmm} function.}

\item{mu}{\code{mu} parameter of the \code{lrpsadmm} function.}

\item{verbose}{A boolean. Whether to print the value of lambda, gamma, sparsity of S and rank of L after each fit.}
}
\value{
An object of class lscggmadmmpath This is essentially a list (see examples). Each element is itself a list with keys:
  \describe{
     \item{lambda}{Value of lambda used for that fit. Recall that the value of the tuning parameters \eqn{\lambda_1, \lambda_2}
     is given by \eqn{\lambda_1 = lambda * gamma} and \eqn{\lambda_2 = lambda * (1 - gamma)}}
     \item{gamma}{Value of gamma used for that fit. }
     \item{lambda1}{Corresponds to the parameter \code{l1} given as argument to the \code{lscggm} function.
     lambda1 = lambda * gamma}
     \item{lambda2}{Corresponds to the parameter \code{l2} given as argument to the \code{lscggm} function.
     lambda2 = lambda * (1 - gamma)}
     \item{number.of.edges}{Number of edges in the estimated sparse graphical model.}
     \item{rank.L}{Rank of the estimated low-rank matrix L.}
     \item{sparsity}{Sparsity of the estimated sparse matrix. This is fraction of entries that are non-zero.}
     \item{fit}{An object of class lscggmadmm
     This is the outcome of calling \code{lscggm} with tuning parameters l1 = lambda * gamma and l2 = lambda (1 - gamma).
     See the documentation of the function \code{lscggm} for more information.}
  }
}
\description{
The penalty for the conditional low-rank plus sparse estimator is written as \eqn{\lambda_1 ||S||_1 + \lambda_2 Trace(L)} in the
objective function of \code{lrpsadmm}.
This can be equivalently rewritten in terms of the regularisation parameters \eqn{\lambda} and \eqn{\gamma} as follows
\deqn{\lambda \gamma ||S||_1 + \lambda (1 - \gamma) Trace(L),}
for \eqn{\gamma \in (0, 1)}.
This function estimates the path of the estimator for a fixed value of \eqn{\gamma} (which controls the trade-off between
the two penalties) by varying the value of \eqn{\lambda}. See the documentation of \code{lrpsadmm} and references therein for
more details.
}
\details{
The function \code{lscggmadmm} is fitted for successive values of \eqn{\lambda} using warm starts.
The sequence of values of \eqn{\lambda} can be provided directly by the user.
It is automatically sorted in decreasing order.
By default, a decreasing sequence of 20 values within a reasonable range is selected as follows.
We set \eqn{\lambda_{max} = \max_{ij, i \neq j} |\SigmaX_{ij}|/\gamma} and
\eqn{\lambda_{min} = \lambda_{max}} * \code{lambda.ratio}; then
20 values between \eqn{\lambda_{max}} and \eqn{\lambda_{min}} are taken following
a geometric progression.

Because it does not make much sense to fit this estimator when the sparse estimate S becomes too dense
or if the rank of the low-rank estimate L becomes too high, the computation of the path is aborted
when the sparsity of S reaches \code{max.sparsity} or when the rank of L reaches \code{max.rank}.
}
\seealso{
lscggm lscggm.cv
}
