% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrps.R
\name{plot.lrpsadmm}
\alias{plot.lrpsadmm}
\title{Plotting function for 'lrpsadmm' Objects}
\usage{
\method{plot}{lrpsadmm}(x)
}
\arguments{
\item{x}{An object of class lrpsadmm output by the function \code{lrpsadmm}}
}
\description{
Plots the sparsity pattern of S, the eigenvalues of L and
the values taken by the objective function at each iteration.
}
\examples{
set.seed(1)
sim.data <- generate.latent.ggm.data(n=2000, p=100, h=5, outlier.fraction = 0.0,
                                    sparsity = 0.02, sparsity.latent = 0.7)
X <- sim.data$obs.data; Sigma <- cor(X) # Sample correlation matrix
lambda <- 0.7; gamma <- 0.1 # The tuning parameters.
l1 <- lambda * gamma; l2 <- lambda * (1 - gamma)
fit <- lrpsadmm(Sigma = Sigma, Lambda1 = l1, Lambda2 = l2)
plot(fit)
}
