% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{plot.lrpsadmmcv}
\alias{plot.lrpsadmmcv}
\title{Plotting for Objects of class "lrpsadmmcv"}
\usage{
\method{plot}{lrpsadmmcv}(x)
}
\arguments{
\item{x}{An object of class lrpsadmmcv output by the \code{lrpsadmm.cv} function.}
}
\description{
Plots an object of class lrpsadmmcv (output by the \code{lrpsadmm.cv} function).
The function plots the support and spectrum of the fit that achieves the best cross-validated negative log-likelihood,
along with the cross-validated log-likelihood (and its standard deviation) for each value of gamma
}
\seealso{
plot.lrpsadmm plot.lrpsadmmpath
}
