% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{show.cross.validate.low.rank.plus.sparse}
\alias{show.cross.validate.low.rank.plus.sparse}
\title{Plot the output of a cross-validated low rank plus sparse model}
\usage{
show.cross.validate.low.rank.plus.sparse(lrps.path, ground.truth = NULL)
}
\description{
Plot the output of a cross-validated low rank plus sparse model
}
