% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_path.R
\name{show.low.rank.plus.sparse.path}
\alias{show.low.rank.plus.sparse.path}
\title{Plot the output fit.low.rank.plus.sparse.path}
\usage{
show.low.rank.plus.sparse.path(lrps.path, ground.truth = NULL)
}
\description{
Plot the output fit.low.rank.plus.sparse.path
}
