% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_estimators.R
\name{Kendall.correlation.estimator}
\alias{Kendall.correlation.estimator}
\title{Estimate the scale matrix of a semiparametric elliptical copula (a.k.a.
transelliptical distribution).}
\usage{
Kendall.correlation.estimator(X)
}
\arguments{
\item{X}{An n x p data matrix.}
}
\value{
A p x p correlation matrix.
}
\description{
Estimate the correlation matrix of a semiparametric elliptical
copula using a modified Kendall rank correlation matrix
}
\details{
Given X, an n x p data matrix, let K be its Kendall correlation matrix.
This function first computes \code{Chat <- sin(0.5*pi*K)} and projects it onto the
space of correlation matrices (in Frobenius norm) using the \code{nearPD} function
of the \code{Matrix} package
}
